requirements.txt
# 手势识别鼠标控制器依赖包
opencv-python>=4.5.0
mediapipe>=0.10.0,<0.11.0
pynput>=1.7.0
pillow>=8.0.0
numpy>=1.20.0